requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
firebase-admin==5.2.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.48.0
grpcio-status==1.48.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
install==1.3.5
Markdown==3.4.1
msgpack==1.0.4
proto-plus==1.20.6
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
zipp==3.8.1
